rename : db 관련 디렉토리 옮김
