+++ app/src/main/java/com/example/mobile_term_project/MainActivity.java
@@ -14,9 +14,9 @@ import android.widget.Toast;
 
 import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.graphics.Insets;
-import androidx.core.view.ViewCompat;
-import androidx.core.view.WindowInsetsCompat;
+
+import com.example.mobile_term_project.db.SQLiteHelper;
+import com.example.mobile_term_project.db.TableInfo;
 
 public class MainActivity extends AppCompatActivity {
