+++ app/src/main/java/com/example/mobile_term_project/SignUp.java
@@ -1,10 +1,8 @@
 package com.example.mobile_term_project;
 
-import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
-import android.database.sqlite.SQLiteDatabase;
 import android.graphics.Color;
 import android.os.Bundle;
 import android.text.Editable;
@@ -14,6 +12,8 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import com.example.mobile_term_project.db.SQLiteHelper;
 
 public class SignUp extends AppCompatActivity {
